perf: common menu
