+++ package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "test": "jest",
     "lint": "tsc --noEmit && eslint '*/**/*.{js,ts,tsx}' --quiet",
-    "lint:fox": "eslint '*/**/*.{js,ts,tsx}' --quiet",
+    "lint:fox": "eslint '*/**/*.{js,ts,tsx}' --quiet --fix",
     "start": "start-storybook -p 6006",
     "build": "tsc",
     "preversion": "yarn build",
