+++ src/stories/index.stories.tsx
@@ -36,16 +36,11 @@ Just an easy to use **Markdown** editor with \`slash commands\``,
 
 export const Emoji = Template.bind({});
 Emoji.args = {
-  defaultValue: `# Notices
+  defaultValue: `# Emoji
 
-There are three types of editable notice blocks that can be used to callout information:
-
-\\
+\
 :1st_place_medal:
 `,
-  // onChange: (v) => {
-  //   console.log(v())
-  // }
 };
 
 export const TemplateDoc = Template.bind({});
